[Add] getContact,getGroup from EventBeans. [Fix] addContact now really private. [Change] getGroup(id) now getGroupID [Change] IReceiver.reply.messageId now using type long.

+++ src/main/java/cc/sfclub/transform/IReceiver.java
@@ -14,5 +14,5 @@ public interface IReceiver {
      * @param messageId target
      * @param message   message
      */
-    void reply(int messageId, String message);
+    void reply(long messageId, String message);
 }
